Move DTO conversion to DataLayer
