+++ Facturation/Models/EntrepriseModel.swift
@@ -98,68 +98,3 @@ final class EntrepriseModel {
     }
 }
 
-// MARK: - Extension de conversion EntrepriseModel <-> EntrepriseDTO
-extension EntrepriseModel {
-    func toDTO() -> EntrepriseDTO {
-        return EntrepriseDTO(
-            id: self.id,
-            nom: self.nom,
-            nomContact: self.nomContact,
-            nomDirigeant: self.nomDirigeant,
-            telephone: self.telephone,
-            email: self.email,
-            siret: self.siret,
-            numeroTVA: self.numeroTVA,
-            iban: self.iban,
-            bic: self.bic,
-            adresseRue: self.adresseRue,
-            adresseCodePostal: self.adresseCodePostal,
-            adresseVille: self.adresseVille,
-            adressePays: self.adressePays,
-            certificationTexte: self.certificationTexte,
-            logo: self.logo,
-            prefixeFacture: self.prefixeFacture,
-            prochainNumero: self.prochainNumero,
-            tvaTauxDefaut: self.tvaTauxDefaut,
-            delaiPaiementDefaut: self.delaiPaiementDefaut,
-            domaine: self.domaine
-        )
-    }
-
-    static func fromDTO(_ dto: EntrepriseDTO) -> EntrepriseModel {
-        let entreprise = EntrepriseModel()
-        entreprise.id = dto.id
-        entreprise.nom = dto.nom
-        entreprise.adresseRue = dto.adresseRue
-        entreprise.adresseCodePostal = dto.adresseCodePostal
-        entreprise.adresseVille = dto.adresseVille
-        entreprise.adressePays = dto.adressePays
-        entreprise.email = dto.email
-        entreprise.telephone = dto.telephone
-        entreprise.siret = dto.siret
-        entreprise.numeroTVA = dto.numeroTVA
-        entreprise.prefixeFacture = dto.prefixeFacture
-        entreprise.iban = dto.iban
-        entreprise.bic = dto.bic
-        entreprise.logo = dto.logo
-        entreprise.domaine = dto.domaine
-        return entreprise
-    }
-
-    func updateFromDTO(_ dto: EntrepriseDTO) {
-        nom = dto.nom
-        adresseRue = dto.adresseRue
-        adresseCodePostal = dto.adresseCodePostal
-        adresseVille = dto.adresseVille
-        adressePays = dto.adressePays
-        email = dto.email
-        telephone = dto.telephone
-        siret = dto.siret
-        numeroTVA = dto.numeroTVA
-        prefixeFacture = dto.prefixeFacture
-        iban = dto.iban
-        bic = dto.bic
-        logo = dto.logo
-        domaine = dto.domaine
-    }
-}
