+++ Facturation/Models/FactureModel.swift
@@ -107,55 +107,4 @@ extension FactureModel {
     }
 }
 
-// MARK: - Extension de conversion FactureModel <-> FactureDTO
-extension FactureModel {
-    func toDTO() -> FactureDTO {
-        return FactureDTO(
-            id: self.id,
-            numero: self.numero,
-            dateFacture: self.dateFacture,
-            dateEcheance: self.dateEcheance,
-            datePaiement: self.datePaiement,
-            tva: self.tva,
-            conditionsPaiement: self.conditionsPaiement.rawValue,
-            remisePourcentage: self.remisePourcentage,
-            statut: self.statut.rawValue,
-            notes: self.notes,
-            notesCommentaireFacture: self.notesCommentaireFacture,
-            clientId: self.client?.id ?? UUID(),
-            ligneIds: self.lignes.map { $0.id }
-        )
-    }
 
-    static func fromDTO(_ dto: FactureDTO, context: ModelContext, client: ClientModel?, lignes: [LigneFacture]) -> FactureModel {
-        let facture = FactureModel()
-        facture.id = dto.id
-        facture.numero = dto.numero
-        facture.dateFacture = dto.dateFacture
-        facture.dateEcheance = dto.dateEcheance
-        facture.datePaiement = dto.datePaiement
-        facture.client = client
-        facture.lignes = lignes
-        facture.tva = dto.tva
-        facture.conditionsPaiement = ConditionsPaiement(rawValue: dto.conditionsPaiement) ?? .virement
-        facture.remisePourcentage = dto.remisePourcentage
-        facture.statut = StatutFacture(rawValue: dto.statut) ?? .brouillon
-        facture.notes = dto.notes
-        facture.notesCommentaireFacture = dto.notesCommentaireFacture
-        context.insert(facture)
-        return facture
-    }
-
-    func updateFromDTO(_ dto: FactureDTO) {
-        numero = dto.numero
-        dateFacture = dto.dateFacture
-        dateEcheance = dto.dateEcheance
-        datePaiement = dto.datePaiement
-        tva = dto.tva
-        conditionsPaiement = ConditionsPaiement(rawValue: dto.conditionsPaiement) ?? .virement
-        remisePourcentage = dto.remisePourcentage
-        statut = StatutFacture(rawValue: dto.statut) ?? .brouillon
-        notes = dto.notes
-        notesCommentaireFacture = dto.notesCommentaireFacture
-    }
-}
